Allow pushing existing code

+++ deployment/push-code.sh
@@ -23,18 +23,23 @@ if [ -z "$GIT_IDENTIFIER" ]; then
   GIT_IDENTIFIER="main"
 fi
 
-echo "Setting up tmp directory"
-TMP_DIRECTORY="/tmp/deploy-checkout"
-rm -rf TMP_DIRECTORY
-mkdir -p $TMP_DIRECTORY
+if [ -z $REPOSITORY_DIRECTORY ]; then
+  echo "Setting up tmp directory"
+  TMP_DIRECTORY="/tmp/deploy-checkout"
+  rm -rf $TMP_DIRECTORY
+  mkdir -p $TMP_DIRECTORY
 
-echo "Cloning repository"
-git clone $GIT_REPO $TMP_DIRECTORY
-cd $TMP_DIRECTORY
+  echo "Cloning repository"
+  git clone $GIT_REPO $TMP_DIRECTORY
+  cd $TMP_DIRECTORY
 
-echo "Pulling target code from repository"
-git fetch
-git checkout $GIT_IDENTIFIER
+  echo "Pulling target code from repository"
+  git fetch
+  git checkout $GIT_IDENTIFIER
+else
+  echo "Git directory supplied. Using that directory"
+  cd $REPOSITORY_DIRECTORY
+fi
 
 echo "Installing dependencies"
 npm install
@@ -43,7 +48,9 @@ echo "Pushing code to remote"
 APP_NAME=$(echo_pm2_app_name $ENV_NAME)
 TARGET_CODE_DIRECTORY="$ENV_DIRECTORY/$APP_NAME"
 ssh -o StrictHostKeyChecking=no -i $KEY_FILE_PATH $SSH_TARGET "rm -rf $TARGET_CODE_DIRECTORY"
-rsync -Pav -e "ssh -i $KEY_FILE_PATH" /tmp/deploy-checkout/* $SSH_TARGET:$TARGET_CODE_DIRECTORY
+rsync -Pav -e "ssh -i $KEY_FILE_PATH" ./* $SSH_TARGET:$TARGET_CODE_DIRECTORY
 
 echo "Cleaning Up"
-rm -rf $TMP_DIRECTORY
+if [ ! -z $TMP_DIRECTORY ]; then
+  rm -rf $TMP_DIRECTORY
+fi
